// scripts/Game/Enums/Upgrades.ts
export enum Upgrades {
    Revealer1 = 1,
    Revealer2 = 2,
    Revealer3 = 3,
    Revealer4 = 4,
    Flagger1 = 5,
    Flagger2 = 6,
    Flagger3 = 7,
    Flagger4 = 8,
    RevealerSpeed1 = 9,
    RevealerSpeed2 = 10,
    RevealerSpeed3 = 11,
    RevealerSpeed4 = 12,
    FlaggerSpeed1 = 13,
    FlaggerSpeed2 = 14,
    FlaggerSpeed3 = 15,
    FlaggerSpeed4 = 16,
}